requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.7.1